+++ server/v2/cometbft/config.go
@@ -3,6 +3,7 @@ package cometbft
 import (
 	cmtcfg "github.com/cometbft/cometbft/config"
 
+	"cosmossdk.io/schema/indexer"
 	"cosmossdk.io/server/v2/cometbft/mempool"
 )
 
@@ -23,6 +24,10 @@ func DefaultAppTomlConfig() *AppTomlConfig {
 		Trace:           false,
 		Standalone:      false,
 		Mempool:         mempool.DefaultConfig(),
+		Indexer: indexer.IndexingConfig{
+			Target:            make(map[string]indexer.Config),
+			ChannelBufferSize: 1024,
+		},
 	}
 }
 
@@ -37,7 +42,8 @@ type AppTomlConfig struct {
 	Standalone      bool     `mapstructure:"standalone" toml:"standalone" comment:"standalone starts the application without the CometBFT node. The node should be started separately."`
 
 	// Sub configs
-	Mempool mempool.Config `mapstructure:"mempool" toml:"mempool" comment:"mempool defines the configuration for the SDK built-in app-side mempool implementations."`
+	Mempool mempool.Config         `mapstructure:"mempool" toml:"mempool" comment:"mempool defines the configuration for the SDK built-in app-side mempool implementations."`
+	Indexer indexer.IndexingConfig `mapstructure:"indexer" toml:"indexer" comment:"indexer defines the configuration for the SDK built-in indexer implementation."`
 }
 
 // CfgOption is a function that allows to overwrite the default server configuration.